python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
ddtrace==2.8.5
datadog==0.49.1
python-dotenv==1.0.1
wrapt>=1.15.0